refactor: renamed it to VectorFilter
