+++ sites/geohub/src/components/controls/VectorFilter.svelte
